site-tools/GitlabMonitor/requirements.txt
python-gitlab==5.6.0
python-dotenv==1.0.1 